selenium-base-docker/requirements.txt
seleniumbase==4.32.2
pyyaml==6.0.2